// DriverProfile/src/model/driver.ts
export interface Driver {
        name: string; 
        executorMemory: string; 
        proxyUser: string; 
        driverCores: number; 
        className: string; 
        driverMemory: string; 
        file: string; 
        type: string; 
        numberofExecutors: number; 
        executorCores: number; 
  }